+++ Vibe/Settings/SettingsAdvancedViewController.h
@@ -1,0 +1,10 @@
+//
+//  SettingsAdvancedViewController.h
+//  Vibe
+//
+
+#import "SettingsPaneViewController.h"
+
+@interface SettingsAdvancedViewController : SettingsPaneViewController
+
+@end