+++ Vibe/Settings/SettingsGeneralViewController.h
@@ -1,0 +1,10 @@
+//
+//  SettingsGeneralViewController.h
+//  Vibe
+//
+
+#import "SettingsPaneViewController.h"
+
+@interface SettingsGeneralViewController : SettingsPaneViewController
+
+@end