+++ Vibe/Settings/SettingsPlaybackViewController.h
@@ -1,0 +1,10 @@
+//
+//  SettingsPlaybackViewController.h
+//  Vibe
+//
+
+#import "SettingsPaneViewController.h"
+
+@interface SettingsPlaybackViewController : SettingsPaneViewController
+
+@end