+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Added settings window
 * Added Czech, Slovak, Hungarian, Greek, Vietnamese, Indonesian, and Thai localizations (now 30 languages)
 * Fixed ⌘W not closing the Settings and About windows